only use characters that are completely random
As the comment describes, the highest order symbols in the base58 string
may not be completely random, since not all values are possible. Let's:
reverse the string, using the lowest order bits first, so we're sure that
they're random, and limit the string to an encoding size that is less than
the size of 2^512, so no not-completely-random symbols can be used.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>